@@ -12,13 +12,15 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # build by specifying --build-name and --build-number, respectively.
 # In Android, build-name is used as versionName while build-number used as versionCode.
 # Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
+# In iOS, build-name is used as CFBundleShortVersionString while build-number is used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
+# In Windows, build-name is used as the major, minor, and patch parts
+# of the product and file versions while build-number is used as the build suffix.
 version: 1.2.1+8
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: '>=3.3.3 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -27,28 +29,12 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
+  cookie_jar: ^4.0.8
+  dio: ^5.4.2+1
+  dio_cookie_manager: ^3.1.1
   flutter:
     sdk: flutter
-  carousel_slider: ^4.1.1
-  dio: ^5.4.1
-  cookie_jar: 4.0.8
-  dio_cookie_manager: ^3.1.0
-  shared_preferences: 2.0.13
-  rxdart: ^0.27.3
-  image_picker: ^0.8.5
-  image_cropper: ^2.0.3
-  # Need to be replace with a maintained dependency
-  date_time_picker:
-    git:
-      url: https://github.com/koai-dev/date_time_picker
-      ref: f9802b8
-  eventify: ^1.0.0
-  fl_chart: ^0.51.0
-  flutter_localizations:
-    sdk: flutter
-  flutter_image_compress: ^1.1.0
-  jwt_decoder: ^2.0.1
-  path_provider: ^2.1.0
+  path_provider: ^2.1.2
 
 dev_dependencies:
   flutter_test:
@@ -59,7 +45,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^1.0.0
+  flutter_lints: ^3.0.0
   flutter_launcher_icons: "^0.9.2"
 
 flutter_icons:
@@ -67,11 +53,13 @@ flutter_icons:
   image_path: "assets/images/ic_launcher.png"
   adaptive_icon_background: "35b4b6"
   adaptive_icon_foreground: assets/images/logo_foreground.png
+  web:
+    generate: true
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
-# The following section is specific to Flutter.
+# The following section is specific to Flutter packages.
 flutter:
 
   # The following line ensures that the Material Icons font is
@@ -84,7 +72,7 @@ flutter:
     - assets/images/placeholder.jpg
 
   # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
+  # https://flutter.dev/assets-and-images/#resolution-aware
 
   # For details regarding adding assets from package dependencies, see
   # https://flutter.dev/assets-and-images/#from-packages
